fix late night typo and address comments
